Temporary if statement to restart DAG task on cloud.

+++ dags/dataset_ETL_OBIS_sequences.py
@@ -15,11 +15,14 @@ from datetime import datetime, timedelta
 csv.field_size_limit(sys.maxsize)
 
 def fetch_OBIS_table():
-    obisdata = occ.search(hasextensions="DNADerivedData", startdepth=200, enddepth=12000).execute()
-
-    obisdata.to_csv('output.csv', index=False)
 
     logging.info(f"Looking for /mnt/bucket/output.csv...")
+
+    if os.path.exists(f"/mnt/bucket/output.csv"):  # temporary because we don't want to wait for the re-do during dev
+        return
+    else:
+        obisdata = occ.search(hasextensions="DNADerivedData", startdepth=200, enddepth=12000).execute()
+        obisdata.to_csv('output.csv', index=False)
 
     if os.path.exists(f"/mnt/bucket/output.csv"):
         logging.info("Download successful!")
@@ -125,7 +128,7 @@ def load_OBIS_table_csv():
             subphylum TEXT,
             subphylumid TEXT,
             taxonRank TEXT,
-            id TEXT PRIMARY KEY,
+            id UUID PRIMARY KEY,
             dataset_id TEXT,
             node_id TEXT,
             dropped TEXT,
